Fixes issue with color setting where the colors array is not properly cast to the a consistent type

+++ masknmf/demixing/demixing_arrays/colorful_ac_array.py
@@ -81,7 +81,7 @@ class ColorfulACArray(FactorizedVideo):
         Args:
             new_colors (torch.tensor): Shape (num_neurons, 3)
         """
-        self._colors = new_colors.to(self.device)
+        self._colors = new_colors.to(self.device).to(self.c.dtype)
 
     @property
     def dtype(self) -> str:
